Initial commit of Clojure Boot script to facilitate file format conversions
Requirements: Clojure Boot, Java 1.8 (or newer)
This script replaces redundant copies of annotations in the
distribution by facilitating dynamic conversion of annotation files
from one format to another at the request of the user. The script is
also capable of creating and populating the directory structure
required for Knowtator-2 annotation projects.