Updated coreference annotations
The coreference annotations have been revised to resolve instances of
identity chains sharing mentions. The original knowtator files have
been removed and replaced with knowtator-2 format files that contain
the revised annotations. For details on the changes to the coreference
annotations, please see this
[README](https://github.com/UCDenver-ccp/CRAFT/blob/master/coreference-a
nnotation/README.md)

Changes to Boot script:
* new task — convert-all: creates all possible file formats for all
annotation types
* new task — check-env: used internally to ensure sufficient JVM heap
allocation prior to running any task using all-concepts
* removed :conll-coref-appos output format — the conllcoref format is
not appropriate for appositive relations as they are not chains
* removed sentence annotation type — sentences are available in the
part-of-speech annotation type
* removed typography annotation type — typography annotations are
available in the document-section annotation type
* new task — clean-coreference-to-knowtator2: used to create the
knowtator-2 format ted files that are part of this commit that contain
updated coreference annotations.