Reverting dependency format from CoNLL-U back to CoNLL-X
Returning the dependency file format for the CRAFT dependency data back
to the CoNLL-X format. The CoNLL-U files were improperly formatted
(XPOS and UPOS columns were mistakenly swapped among other things) and
there is no UPOS data to include. This change aligns more closely with
the original CRAFT dependency files (available in CRAFT v3.0 and
earlier releases). Those original files were missing one POS column
which is now included to fully comply with the CoNLL-X format. The
non-compliant CoNLL-U files have been removed from the distribution.
Also, a minor change was made to the HeadRule used in the conversion
from treebank files to the dependency files. The CONLL HeadRule is now
used instead of the STANFORD HeadRule.

Update of file-conversion dependency to 0.2.1 includes changes for the
CoNLL-X file generation as well as updates to allow the boot script to
work with Java >= 9.0.