re-encode CL.obo.zip and CL+extensions.obo.zip from Mac-Roman to UTF-8
Except for CL, all ontologies packaged in CRAFT are encoded with UTF-8 (or its subset ASCII), so it seems natural to use that codec for CL too.
The official file provided by the OBO foundry is encoded with UTF-8 as well.

Besides a single occurrence of "ü" in "Müller cell", all non-ASCII characters are (in part doubly-encoded) instance of the non-break space character (U+A0).
Since these characters are somewhat hidden, are rarely used otherwise in the ontology and don't seem to serve an obvious purpose here, they are replaced with a regular space in this commit.
This affects the definitions of CL:0002323 and CL:0002363.
The changes are exactly the same for the plain and the extended ontology.